feat: system status API, light theme config, secret-safe gitignore
- 新增 /api/system/status 端点和首页系统状态展示
- 新增 .env.example 配置模板
- 新增 test_config.py / test_system_status.py 测试
- 更新 .gitignore 排除 .artifacts/ 和 frontend-prototype 构建产物
- 修复 TemplateResponse 兼容性（6 处）
- 统一 AI provider 配置：AIClient 支持多 provider 自动推断
- 验证结果：67 passed

+++ app/models.py
@@ -1,5 +1,5 @@
 import datetime
-from sqlalchemy import Column, Integer, String, Text, DateTime, DECIMAL, ForeignKey
+from sqlalchemy import Column, Integer, String, Text, DateTime, Float, ForeignKey
 from app.database import Base
 
 
@@ -58,7 +58,7 @@ class PRReviewFinding(Base):
     title = Column(String(512), nullable=True)
     reason = Column(Text, nullable=True)
     suggestion = Column(Text, nullable=True)
-    confidence = Column(DECIMAL(4, 3), nullable=True)
+    confidence = Column(Float, nullable=True)
     evidence_json = Column(Text, nullable=True)
     status = Column(String(32), default="OPEN")
     created_at = Column(DateTime, default=datetime.datetime.utcnow)
